+++ src/registry/docker/docker.go
