+++ src/registry/quay/quay.go
