+++ src/rollout.go
@@ -2,12 +2,13 @@ package reroller
 
 import (
 	"context"
+	"time"
+
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/kubernetes"
-	"time"
 )
 
 const restartedAtAnnotation = "reroller.roob.re/restartedAt"
